Move presets to shared file and add api endpoint
