+++ packages/client/index.html
@@ -77,6 +77,7 @@
     <!-- IMAGE_VERSION -->
     <!-- GIT_COMMIT -->
     <!-- USER_ID -->
+    <!-- SYSTEM_PRESETS -->
     <script>
       function tabControl(evt, cityName) {
         // Declare all variables
