+++ packages/server/src/secret.ts
@@ -1,12 +1,15 @@
 import fs from 'fs-extra';
 
 let jwtSecret: string = '';
+let apiSecret: string = '';
 
 function initSecrets(): void {
   if (process.env.NODE_ENV === 'production') {
     jwtSecret = fs.readFileSync('/run/secrets/jwt_secret', 'utf8').trim();
+    jwtSecret = fs.readFileSync('/run/secrets/api_secret', 'utf8').trim();
   } else {
     jwtSecret = 'example_secret_key';
+    apiSecret = 'example_secret_key';
   }
 }
 
@@ -14,4 +17,8 @@ function getJwtSecret(): string {
   return jwtSecret;
 }
 
-export { initSecrets, getJwtSecret };
+function getApiSecret(): string {
+  return apiSecret;
+}
+
+export { initSecrets, getJwtSecret, getApiSecret };
