+++ packages/server/src/util/escapeHtml.ts
@@ -1,0 +1,17 @@
+type HtmlCharMap = {
+  [key: string]: string;
+};
+
+const abc: HtmlCharMap = {
+  '&': '&amp;',
+  '<': '&lt;',
+  '>': '&gt;',
+  "'": '&#39;',
+  '"': '&quot;',
+};
+
+export function escapeHtml(str: string) {
+  return str.replace(/[&<>'"]/g, (tag: string) => {
+    return abc[tag] || '';
+  });
+}
